refactor(#792): split i18n_tests into base + catalog-consistency suite (sweep)
i18n_tests.rs 1007 -> 447, + i18n_catalog_tests.rs 551 (both under cap; off debt).
The heavy Slint @tr-key ↔ .po catalog-scanning tests (self-contained, own nested
helpers) move to the sibling. adapter-gui lib tests (540) pass, zero warnings.
